+++ common/src/main/java/za/co/neroland/nerospace/rocket/RocketEntity.java
@@ -899,7 +899,7 @@ public class RocketEntity extends Entity implements MenuProvider {
             za.co.neroland.nerospace.network.ModNetwork.sendToPlayer(serverPlayer,
                     za.co.neroland.nerospace.network.StationSyncPayload.of(
                             StationRegistry.get(serverPlayer.level().getServer())));
-            MenuOpener.open(serverPlayer, this);
+            return MenuOpener.openOrConsume(serverPlayer, this);
         }
         return InteractionResult.SUCCESS;
     }